cleanup

+++ wp-content/plugins/seo-by-rank-math/includes/modules/analytics/class-db.php
@@ -12,8 +12,10 @@ namespace RankMath\Analytics;
 
 use RankMath\Helper;
 use RankMath\Google\Api;
-use MyThemeShop\Helpers\Str;
-use MyThemeShop\Database\Database;
+use RankMath\Google\Console;
+use RankMath\Helpers\Str;
+use RankMath\Helpers\DB as DB_Helper;
+use RankMath\Admin\Database\Database;
 
 defined( 'ABSPATH' ) || exit;
 
@@ -27,16 +29,16 @@ class DB {
 	 *
 	 * @param string $table_name Table name.
 	 *
-	 * @return \MyThemeShop\Database\Query_Builder
+	 * @return \RankMath\Admin\Database\Query_Builder
 	 */
 	public static function table( $table_name ) {
 		return Database::table( $table_name );
 	}
 
 	/**
-	 * Get analytics table.
+	 * Get console data table.
 	 *
-	 * @return \MyThemeShop\Database\Query_Builder
+	 * @return \RankMath\Admin\Database\Query_Builder
 	 */
 	public static function analytics() {
 		return Database::table( 'rank_math_analytics_gsc' );
@@ -45,10 +47,19 @@ class DB {
 	/**
 	 * Get objects table.
 	 *
-	 * @return \MyThemeShop\Database\Query_Builder
+	 * @return \RankMath\Admin\Database\Query_Builder
 	 */
 	public static function objects() {
 		return Database::table( 'rank_math_analytics_objects' );
+	}
+
+	/**
+	 * Get inspections table.
+	 *
+	 * @return \RankMath\Admin\Database\Query_Builder
+	 */
+	public static function inspections() {
+		return Database::table( 'rank_math_analytics_inspections' );
 	}
 
 	/**
@@ -57,15 +68,19 @@ class DB {
 	 * @param  int $days Decide whether to delete all or delete 90 days data.
 	 */
 	public static function delete_by_days( $days ) {
-		if ( -1 === $days ) {
-			self::analytics()->truncate();
-		} else {
-			$start = date_i18n( 'Y-m-d H:i:s', strtotime( '-1 days' ) );
-			$end   = date_i18n( 'Y-m-d H:i:s', strtotime( '-' . $days . ' days' ) );
+		// Delete console data.
+		if ( Console::is_console_connected() ) {
+			if ( -1 === $days ) {
+				self::analytics()->truncate();
+			} else {
+				$start = date_i18n( 'Y-m-d H:i:s', strtotime( '-1 days' ) );
+				$end   = date_i18n( 'Y-m-d H:i:s', strtotime( '-' . $days . ' days' ) );
 
-			self::analytics()->whereBetween( 'created', [ $end, $start ] )->delete();
+				self::analytics()->whereBetween( 'created', [ $end, $start ] )->delete();
+			}
 		}
 
+		// Delete analytics, adsense data.
 		do_action( 'rank_math/analytics/delete_by_days', $days );
 		self::purge_cache();
 
@@ -78,10 +93,12 @@ class DB {
 	public static function delete_data_log() {
 		$days = Helper::get_settings( 'general.console_caching_control', 90 );
 
+		// Delete old console data more than 2 times ago of specified number of days to keep the data.
 		$start = date_i18n( 'Y-m-d H:i:s', strtotime( '-' . ( $days * 2 ) . ' days' ) );
 
 		self::analytics()->where( 'created', '<', $start )->delete();
 
+		// Delete old analytics and adsense data.
 		do_action( 'rank_math/analytics/delete_data_log', $start );
 	}
 
@@ -113,6 +130,10 @@ class DB {
 			return [];
 		}
 
+		if ( ! DB_Helper::check_table_exists( 'rank_math_analytics_gsc' ) ) {
+			return [];
+		}
+
 		$key  = 'rank_math_analytics_data_info';
 		$data = get_transient( $key );
 		if ( false !== $data ) {
@@ -127,7 +148,7 @@ class DB {
 			->selectCount( 'id' )
 			->getVar();
 
-		$size = $wpdb->get_var( 'SELECT SUM((data_length + index_length)) AS size FROM information_schema.TABLES WHERE table_schema="' . $wpdb->dbname . '" AND (table_name="' . $wpdb->prefix . 'rank_math_analytics_gsc")' ); // phpcs:ignore
+		$size = DB_Helper::get_var( "SELECT SUM((data_length + index_length)) AS size FROM information_schema.TABLES WHERE table_schema='" . $wpdb->dbname . "' AND (table_name='" . $wpdb->prefix . "rank_math_analytics_gsc')" );
 		$data = compact( 'days', 'rows', 'size' );
 
 		$data = apply_filters( 'rank_math/analytics/analytics_tables_info', $data );
@@ -158,55 +179,36 @@ class DB {
 	}
 
 	/**
-	 * Check if a date exists in the sysyem.
+	 * Check if console data exists at specified date.
 	 *
-	 * @param  string $date   Date.
-	 * @param  string $action Action.
+	 * @param  string $date   Date to check data existence.
+	 * @param  string $action Action name to filter data type.
 	 * @return boolean
 	 */
 	public static function date_exists( $date, $action = 'console' ) {
-		$table = [
-			'console'   => 'rank_math_analytics_gsc',
-			'analytics' => 'rank_math_analytics_ga',
-			'adsense'   => 'rank_math_analytics_adsense',
-		];
+		$tables['console'] = DB_Helper::check_table_exists( 'rank_math_analytics_gsc' ) ? 'rank_math_analytics_gsc' : '';
 
-		$table = self::table( $table[ $action ] );
+		/**
+		 * Filter: 'rank_math/analytics/date_exists_tables' - Allow developers to add more tables to check.
+		 */
+		$tables = apply_filters( 'rank_math/analytics/date_exists_tables', $tables, $date, $action );
+
+		if ( empty( $tables[ $action ] ) ) {
+			return true; // Should return true to avoid further data fetch action.
+		}
+
+		$table = self::table( $tables[ $action ] );
 
 		$id = $table
 			->select( 'id' )
-			->where( 'created', $date )
+			->where( 'DATE(created)', $date )
 			->getVar();
 
 		return $id > 0 ? true : false;
 	}
 
 	/**
-	 * Check if a date exists in the sysyem.
-	 *
-	 * @param  string $date  Date.
-	 * @param  string $table Table name.
-	 * @return boolean
-	 */
-	public static function job_date_exists( $date, $table = 'console' ) {
-		$tables = [
-			'adsense'   => 'rank_math_analytics_adsense',
-			'analytics' => 'rank_math_analytics_ga',
-			'console'   => 'rank_math_analytics_gsc',
-		];
-		$table  = isset( $tables[ $table ] ) ? $tables [ $table ] : $table;
-		$id     = self::table( $table );
-
-		$id = $id
-			->select( 'id' )
-			->where( 'created', $date )
-			->getVar();
-
-		return $id > 0 ? true : false;
-	}
-
-	/**
-	 * Add a new record.
+	 * Add a new record into objects table.
 	 *
 	 * @param array $args Values to insert.
 	 *
@@ -239,7 +241,76 @@ class DB {
 	}
 
 	/**
-	 * Update a record.
+	 * Add new record in the inspections table.
+	 *
+	 * @param array $args Values to insert.
+	 *
+	 * @return bool|int
+	 */
+	public static function store_inspection( $args = [] ) {
+		if ( empty( $args ) || empty( $args['page'] ) ) {
+			return false;
+		}
+
+		unset( $args['id'] );
+
+		$defaults = self::get_inspection_defaults();
+
+		// Only keep $args items that are in $defaults.
+		$args = array_intersect_key( $args, $defaults );
+
+		// Apply defaults.
+		$args = wp_parse_args( $args, $defaults );
+
+		// We only have strings: placeholders will be '%s'.
+		$format = array_fill( 0, count( $args ), '%s' );
+
+		// Check if we have an existing record, based on 'page'.
+		$id = self::inspections()
+			->select( 'id' )
+			->where( 'page', $args['page'] )
+			->getVar();
+
+		if ( $id ) {
+			return self::inspections()
+				->set( $args )
+				->where( 'id', $id )
+				->update();
+		}
+
+		return self::inspections()->insert( $args, $format );
+	}
+
+	/**
+	 * Get inspection defaults.
+	 *
+	 * @return array
+	 */
+	public static function get_inspection_defaults() {
+		$defaults = [
+			'created'              => current_time( 'mysql' ),
+			'page'                 => '',
+			'index_verdict'        => 'VERDICT_UNSPECIFIED',
+			'indexing_state'       => 'INDEXING_STATE_UNSPECIFIED',
+			'coverage_state'       => '',
+			'page_fetch_state'     => 'PAGE_FETCH_STATE_UNSPECIFIED',
+			'robots_txt_state'     => 'ROBOTS_TXT_STATE_UNSPECIFIED',
+			'rich_results_verdict' => 'VERDICT_UNSPECIFIED',
+			'rich_results_items'   => '',
+			'last_crawl_time'      => '',
+			'crawled_as'           => 'CRAWLING_USER_AGENT_UNSPECIFIED',
+			'google_canonical'     => '',
+			'user_canonical'       => '',
+			'sitemap'              => '',
+			'referring_urls'       => '',
+			'raw_api_response'     => '',
+		];
+
+		return apply_filters( 'rank_math/analytics/inspection_defaults', $defaults );
+	}
+
+	/**
+	 * Add/Update a record into/from objects table.
 	 *
 	 * @param array $args Values to update.
 	 *
@@ -250,6 +321,7 @@ class DB {
 			return false;
 		}
 
+		// If object exists, try to update.
 		$old_id = absint( $args['id'] );
 		if ( ! empty( $old_id ) ) {
 			unset( $args['id'] );
@@ -262,8 +334,17 @@ class DB {
 			if ( ! empty( $updated ) ) {
 				return $old_id;
 			}
+			$old_id = self::objects()
+			->select( 'id' )
+			->where( 'object_id', absint( $args['object_id'] ) )
+			->getVar();
+			if ( ! empty( $old_id ) ) {
+				// $updated may sometimes return 0 if there is no field that is changed, even if a row with $args['object_id'] exists.
+				return $old_id;
+			}
 		}
 
+		// In case of new object or failed to update, try to add.
 		return self::add_object( $args );
 	}
 
@@ -282,7 +363,7 @@ class DB {
 	}
 
 	/**
-	 * Bulk inserts records into a table using WPDB.  All rows must contain the same keys.
+	 * Bulk inserts records into a console table using WPDB.  All rows must contain the same keys.
 	 *
 	 * @param  string $date        Date.
 	 * @param  array  $rows        Rows to insert.
@@ -326,7 +407,7 @@ class DB {
 
 			$data[] = $date;
 			$data[] = $row['query'];
-			$data[] = Stats::get_relative_url( self::remove_hash( $row['page'] ) );
+			$data[] = self::get_page( $row['page'] );
 			$data[] = $row['clicks'];
 			$data[] = $row['impressions'];
 			$data[] = $row['position'];
@@ -344,11 +425,45 @@ class DB {
 		$sql .= implode( ",\n", $placeholders );
 
 		// Run the query.  Returns number of affected rows.
-		return $wpdb->query( $wpdb->prepare( $sql, $data ) ); // phpcs:ignore
+		return DB_Helper::query( $wpdb->prepare( $sql, $data ) );
 	}
 
 	/**
-	 * Remove hash part.
+	 * Get page slug from full URL.
+	 *
+	 * @param  string $url Full URL to parse.
+	 * @return string Page path/slug with leading slash.
+	 */
+	public static function get_page( $url ) {
+		if ( empty( $url ) || ! is_string( $url ) ) {
+			return '';
+		}
+
+		$url = urldecode( preg_replace( '/#.*$/', '', $url ) );
+
+		$url = self::remove_hash( $url );
+
+		// Parse the URL to get the path component.
+		$parsed_url = wp_parse_url( $url );
+		if ( isset( $parsed_url['path'] ) ) {
+			return $parsed_url['path'];
+		}
+
+		// Fallback: try to extract path by removing domain.
+		$host = Helper::get_home_url();
+		$url  = str_replace( $host, '', $url );
+
+		// Remove ASCII domain.
+		$host_ascii = idn_to_ascii( $host );
+		$url        = str_replace( $host_ascii, '', $url );
+
+		$url = preg_replace( '#^https?://(www\.)?#i', '', $url );
+
+		return $url;
+	}
+
+	/**
+	 * Remove hash part from Url.
 	 *
 	 * @param  string $url Url to process.
 	 * @return string
@@ -374,5 +489,53 @@ class DB {
 		}
 
 		return $number;
+	}
+
+	/**
+	 * Get all inspections.
+	 *
+	 * @param array $params   REST Parameters.
+	 * @param int   $per_page Limit.
+	 */
+	public static function get_inspections( $params, $per_page ) {
+		$page     = ! empty( $params['page'] ) ? absint( $params['page'] ) : 1;
+		$per_page = absint( $per_page );
+		$offset   = ( $page - 1 ) * $per_page;
+
+		$inspections = self::inspections()->table;
+		$objects     = self::objects()->table;
+
+		$query = self::inspections()
+			->select( [ "$inspections.*", "$objects.title", "$objects.object_id" ] )
+			->leftJoin( $objects, "$inspections.page", "$objects.page" )
+			->where( "$objects.page", '!=', '' )
+			->orderBy( 'id', 'DESC' )
+			->limit( $per_page, $offset );
+
+		do_action_ref_array( 'rank_math/analytics/get_inspections_query', [ &$query, $params ] );
+
+		$results = $query->get();
+
+		return apply_filters( 'rank_math/analytics/get_inspections_results', $results );
+	}
+
+	/**
+	 * Get inspections count.
+	 *
+	 * @param array $params   REST Parameters.
+	 *
+	 * @return int
+	 */
+	public static function get_inspections_count( $params ) {
+		$inspections = self::inspections()->table;
+		$objects     = self::objects()->table;
+		$query       = self::inspections()
+		->selectCount( "$inspections.id", 'total' )
+		->leftJoin( $objects, "$inspections.page", "$objects.page" )
+		->where( "$objects.page", '!=', '' );
+
+		do_action_ref_array( 'rank_math/analytics/get_inspections_count_query', [ &$query, $params ] );
+
+		return $query->getVar();
 	}
 }
